refactor(dialogue_manager): split tests into per-axis files
Step 9/9 (final). Move the ~37 inline tests in mod.rs across three
sibling files inside dialogue_manager/tests/:

- lifecycle_tests.rs (412 lines, 12 tests): open_session,
  resume_session_*, close_session_* (incl. close_session_drains_*
  and close_session_always_saves_learner_*)
- turn_tests.rs (478 lines, 15 tests): respond_to_streaming_* +
  per_turn_save_* + summary_does_not_refresh_* + the
  respond_to_thin_wrapper test
- background_tests.rs (761 lines, 14 tests): classifier spawn-and-
  await + post-response chain (extract_*, pending_extraction_*,
  post_response_chain_*) + multi-turn end-to-end routing tests

Promotes three per-test mocks/helpers to test_support.rs as
pub(super) since they're now used across files:
- RepeatingBackend (used by per_turn_save_* in turn_tests +
  end_to_end_classifier_routing in background +
  close_session_always_saves_learner in lifecycle)
- ConceptCapturingStore (used by extract_* + post_response_chain_*
  in background + close_session_drains_extractor_task in lifecycle)
- dirty_flag_test_setup (used by per_turn_save_* in turn_tests +
  close_session_always_saves_learner in lifecycle)

FailingLearnerStore stays inline in turn_tests.rs (single-use).

mod.rs: 2002 → 212 lines (89% reduction). The mod tests block now
contains only three submodule declarations.

test_support.rs: 328 → 503 lines (absorbed the three shared mocks).
Still under the spec's documented 700-line soft cap.

background_tests.rs at 761 lines is the only file over the 500-line
soft limit. Splitting classifier-tests from post-response-tests
would be artificial — both exercise the same spawn-and-await pattern
on the dialogue manager. Documented as the second acknowledged
violation alongside test_support.rs.

All 434 tests still pass.
