+++ src/crates/primer-pedagogy/src/dialogue_manager/mod.rs
@@ -206,1797 +206,7 @@ type PostResponseOutcome = Option<
 
 #[cfg(test)]
 mod tests {
-    use super::test_support::*;
-    use super::*;
-    use async_trait::async_trait;
-    use chrono::Utc;
-    use futures::stream;
-    use primer_core::config::PedagogyConfig;
-    use primer_core::conversation::{Speaker, Turn};
-    use primer_core::error::{PrimerError, Result};
-    use primer_core::inference::{
-        GenerationParams, InferenceBackend, Prompt, TokenChunk, TokenStream,
-    };
-    use primer_core::knowledge::KnowledgeBase;
-    use primer_core::learner::{EngagementState, LearnerModel};
-    use primer_extractor::ExtractorSettings;
-    use std::sync::Mutex;
-    use uuid::Uuid;
-
-    #[tokio::test]
-    async fn respond_to_streaming_invokes_callback_per_chunk() {
-        let backend = ScriptedBackend::new(vec![
-            Ok(chunk("Hel", false)),
-            Ok(chunk("lo", false)),
-            Ok(chunk(" there", false)),
-            Ok(chunk("", true)),
-        ]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let received: Mutex<Vec<String>> = Mutex::new(vec![]);
-        let _ = dm
-            .respond_to_streaming("why is the sky blue", |c| {
-                received.lock().unwrap().push(c.to_string());
-            })
-            .await
-            .unwrap();
-
-        let joined: String = received.lock().unwrap().join("");
-        assert_eq!(joined, "Hello there");
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_returns_full_accumulated_text() {
-        let backend = ScriptedBackend::new(vec![
-            Ok(chunk("Hel", false)),
-            Ok(chunk("lo", false)),
-            Ok(chunk(" there", false)),
-            Ok(chunk("", true)),
-        ]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let result = dm.respond_to_streaming("hi", |_| {}).await.unwrap();
-        assert_eq!(result, "Hello there");
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_records_full_primer_turn() {
-        let backend = ScriptedBackend::new(vec![
-            Ok(chunk("part one ", false)),
-            Ok(chunk("part two", false)),
-            Ok(chunk("", true)),
-        ]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.respond_to_streaming("question", |_| {}).await.unwrap();
-        let last = dm.session.turns.last().unwrap();
-        assert_eq!(last.speaker, Speaker::Primer);
-        assert_eq!(last.text, "part one part two");
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_does_not_record_primer_turn_on_stream_error() {
-        let backend = ScriptedBackend::new(vec![
-            Ok(chunk("partial", false)),
-            Err(PrimerError::Inference("simulated network drop".into())),
-        ]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let result = dm.respond_to_streaming("question", |_| {}).await;
-        assert!(result.is_err(), "expected Err on mid-stream failure");
-
-        // Child turn should be recorded; Primer turn should NOT be.
-        assert_eq!(dm.session.turns.len(), 1);
-        assert_eq!(dm.session.turns[0].speaker, Speaker::Child);
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_preserves_typed_inference_error_variant() {
-        // Regression test for the dialogue_manager.rs:534 fix (commit c1578251).
-        // Before the fix, a .map_err wrap re-wrapped typed InferenceError
-        // variants from the backend back into InferenceError::Other via
-        // format!(...).into(). That destroyed the typed dispatch the i18n
-        // render layer relies on — a 401 from Anthropic landed as
-        // Other("Generation failed: ...") and the user saw "Something
-        // unexpected went wrong" instead of the friendly Auth message.
-        //
-        // This test asserts that a typed Auth variant from the backend
-        // survives the dialogue_manager round-trip with its variant intact.
-        let backend = ScriptedBackend::new(vec![Err(PrimerError::Inference(
-            primer_core::error::InferenceError::Auth,
-        ))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        let result = dm.respond_to_streaming("question", |_| {}).await;
-        assert!(
-            matches!(
-                result,
-                Err(PrimerError::Inference(
-                    primer_core::error::InferenceError::Auth
-                ))
-            ),
-            "expected typed Auth variant to survive round-trip, got: {result:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_returns_empty_string_when_stream_yields_no_text() {
-        // Backend completes cleanly with only an empty done-chunk. The call
-        // should succeed with an empty accumulated string and still record
-        // the (empty) Primer turn — the consumer is responsible for noticing
-        // and surfacing this as a user-facing problem if they care.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let result = dm.respond_to_streaming("hi", |_| {}).await.unwrap();
-        assert_eq!(result, "");
-        let last = dm.session.turns.last().unwrap();
-        assert_eq!(last.speaker, Speaker::Primer);
-        assert_eq!(last.text, "");
-    }
-
-    #[tokio::test]
-    async fn respond_to_thin_wrapper_still_works() {
-        let backend = ScriptedBackend::new(vec![
-            Ok(chunk("alpha ", false)),
-            Ok(chunk("beta", false)),
-            Ok(chunk("", true)),
-        ]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let result = dm.respond_to("hi").await.unwrap();
-        assert_eq!(result, "alpha beta");
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_fires_engine_save_on_success() {
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hi", false)), Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let store = Arc::new(CountingStore::new());
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&store) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.respond_to_streaming("hello", |_| {}).await.unwrap();
-
-        // Engine fired exactly one save. Persisted session has both the
-        // child input and the Primer response.
-        assert_eq!(store.save_count(), 1);
-        assert_eq!(store.last_turn_count(), 2);
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_fires_engine_save_on_stream_error() {
-        let backend = ScriptedBackend::new(vec![
-            Ok(chunk("partial", false)),
-            Err(PrimerError::Inference("simulated drop".into())),
-        ]);
-        let knowledge = EmptyKnowledge;
-        let store = Arc::new(CountingStore::new());
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&store) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let result = dm.respond_to_streaming("question", |_| {}).await;
-        assert!(result.is_err());
-
-        // Engine fired the save even though the stream errored. Persisted
-        // session has only the child turn (Primer turn was dropped).
-        assert_eq!(store.save_count(), 1);
-        assert_eq!(store.last_turn_count(), 1);
-        assert_eq!(dm.session.turns.len(), 1);
-        assert_eq!(dm.session.turns[0].speaker, Speaker::Child);
-    }
-
-    #[tokio::test]
-    async fn close_session_fires_engine_save_with_ended_at() {
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hi", false)), Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let store = Arc::new(CountingStore::new());
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&store) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.respond_to_streaming("hello", |_| {}).await.unwrap();
-        // First save fired during respond_to_streaming.
-        let saves_after_response = store.save_count();
-
-        dm.close_session().await;
-
-        // close_session also fires a save, this time with ended_at populated.
-        assert_eq!(store.save_count(), saves_after_response + 1);
-        assert!(dm.session.ended_at.is_some());
-    }
-
-    #[tokio::test]
-    async fn open_session_fires_engine_save() {
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let store = Arc::new(CountingStore::new());
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&store) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.open_session().await.unwrap();
-
-        // The greeting turn was recorded and persisted.
-        assert_eq!(store.save_count(), 1);
-        assert_eq!(store.last_turn_count(), 1);
-    }
-
-    // ─── resume_session and summary refresh ──────────────────────────
-
-    #[tokio::test]
-    async fn resume_session_loads_turns_without_greeting() {
-        // Resume picks up the loaded turns verbatim. No greeting is
-        // prepended; the turn count after resume_session matches the
-        // loaded session exactly.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        let learner_id = dm.learner.profile.id;
-        let loaded = make_test_session_with_turns(5, learner_id);
-        let loaded_id = loaded.id;
-        dm.resume_session(loaded).await.unwrap();
-        assert_eq!(dm.session.turns.len(), 5);
-        assert_eq!(dm.session.id, loaded_id);
-        // The Primer never said "Hello, ..." — turn 0 is from our test fixture.
-        assert_eq!(dm.session.turns[0].text, "turn 0");
-    }
-
-    #[tokio::test]
-    async fn resume_session_clears_ended_at_and_persists() {
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let store = Arc::new(CountingStore::new());
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&store) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        let mut loaded = make_test_session_with_turns(3, dm.learner.profile.id);
-        loaded.ended_at = Some(Utc::now());
-        dm.resume_session(loaded).await.unwrap();
-        assert!(dm.session.ended_at.is_none(), "ended_at should be cleared");
-        assert_eq!(store.save_count(), 1, "resume should fire one save");
-    }
-
-    #[tokio::test]
-    async fn resume_session_preserves_loaded_learner_id() {
-        // The in-memory LearnerModel comes from CLI flags; the loaded
-        // Session might belong to a different learner_id. Resume must
-        // keep the loaded learner_id (no silent override).
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        let dm_learner_id = dm.learner.profile.id;
-        let other_learner = Uuid::new_v4();
-        assert_ne!(dm_learner_id, other_learner);
-        let loaded = make_test_session_with_turns(2, other_learner);
-        dm.resume_session(loaded).await.unwrap();
-        assert_eq!(
-            dm.session.learner_id, other_learner,
-            "session learner_id should not be overwritten by the manager's learner"
-        );
-    }
-
-    #[tokio::test]
-    async fn resume_session_triggers_summary_refresh_when_above_window() {
-        // A loaded session with > context_window_turns should get its
-        // summary refreshed unconditionally on resume so the Primer has
-        // long-term memory of pre-window turns from turn one.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        // Window is 20; 25 turns gives 5 pre-window turns.
-        let loaded = make_test_session_with_turns(25, dm.learner.profile.id);
-        dm.resume_session(loaded).await.unwrap();
-        assert_eq!(
-            backend.summary_call_count(),
-            1,
-            "summary should refresh on resume"
-        );
-        assert!(
-            !dm.session.summary.is_empty(),
-            "summary should be populated after refresh"
-        );
-        assert_eq!(
-            dm.session.summary_through_turn_index, 5,
-            "summary boundary should land at total - window"
-        );
-    }
-
-    #[tokio::test]
-    async fn resume_session_skips_summary_when_inside_first_window() {
-        // Sessions that fit inside the active window have nothing to
-        // summarize; resume must not waste an inference call.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        // Window is 20; 5 turns is well inside.
-        let loaded = make_test_session_with_turns(5, dm.learner.profile.id);
-        dm.resume_session(loaded).await.unwrap();
-        assert_eq!(backend.summary_call_count(), 0);
-        assert_eq!(dm.session.summary, "");
-    }
-
-    #[tokio::test]
-    async fn resume_session_skips_refresh_when_summary_already_current() {
-        // Loaded session has 25 turns and a summary that already covers
-        // turns[..5] — exactly the pre-window range. There is no new
-        // pre-window content for the summary to absorb, so resume must
-        // not burn an LLM call regenerating identical work.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        let mut loaded = make_test_session_with_turns(25, dm.learner.profile.id);
-        loaded.summary = "Pre-existing summary covering turns 0..5.".to_string();
-        loaded.summary_through_turn_index = 5;
-        dm.resume_session(loaded).await.unwrap();
-        assert_eq!(
-            backend.summary_call_count(),
-            0,
-            "summary already covers the pre-window range; resume must not regenerate"
-        );
-        assert_eq!(
-            dm.session.summary, "Pre-existing summary covering turns 0..5.",
-            "existing summary must be preserved verbatim"
-        );
-    }
-
-    #[tokio::test]
-    async fn resume_session_refreshes_when_existing_summary_is_stale() {
-        // Loaded session has 30 turns and a summary that only covers
-        // turns[..3]. The current pre-window range is turns[..10], so
-        // there are 7 pre-window turns the summary doesn't yet know
-        // about. Resume must refresh.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        let mut loaded = make_test_session_with_turns(30, dm.learner.profile.id);
-        loaded.summary = "Stale summary covering only turns 0..3.".to_string();
-        loaded.summary_through_turn_index = 3;
-        dm.resume_session(loaded).await.unwrap();
-        assert_eq!(backend.summary_call_count(), 1);
-        assert_eq!(dm.session.summary_through_turn_index, 10);
-    }
-
-    #[tokio::test]
-    async fn summary_does_not_refresh_when_below_threshold_during_active_session() {
-        // First respond_to_streaming fires only when there are turns to
-        // process. With turn count below window+window, no refresh.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("ok", false)), Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-        // Pre-load with 21 turns (1 turn pre-window). Far below the
-        // 2*window threshold.
-        dm.session.turns = make_test_session_with_turns(21, dm.learner.profile.id).turns;
-        let _ = dm.respond_to_streaming("hi", |_| {}).await.unwrap();
-        assert_eq!(backend.summary_call_count(), 0);
-    }
-
-    // ─── Integration: classifier spawned and applied across turns ─────
-
-    #[tokio::test]
-    async fn resume_session_rehydrates_recent_assessments() {
-        use primer_classifier::{EngagementClassifier, StubEngagementClassifier};
-        use primer_core::classifier::EngagementAssessment;
-        use primer_core::storage::SessionStore;
-        use primer_storage::SqliteSessionStore;
-
-        let storage: Arc<dyn SessionStore> = Arc::new(
-            SqliteSessionStore::open_for_locale(
-                std::path::Path::new(":memory:"),
-                primer_core::i18n::Locale::default(),
-            )
-            .unwrap(),
-        );
-        let classifier: Arc<dyn EngagementClassifier> = Arc::new(StubEngagementClassifier::new());
-
-        // Pre-seed: save a session with one child turn and one classification.
-        let learner = test_learner();
-        let mut session = Session::new(learner.profile.id);
-        session.add_turn(Turn {
-            speaker: Speaker::Child,
-            text: "x".into(),
-            timestamp: Utc::now(),
-            intent: None,
-            concepts: vec![],
-        });
-        storage.save_session(&session).await.unwrap();
-        storage
-            .save_classification(
-                session.id,
-                0,
-                &EngagementAssessment {
-                    state: EngagementState::FrustratedTrying,
-                    confidence: 0.9,
-                    reasoning: Some("test".into()),
-                },
-                "stub",
-            )
-            .await
-            .unwrap();
-
-        // Create a DialogueManager and resume the persisted session.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let settings = ClassifierSettings::default();
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&storage) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            DialogueManagerSubsystems {
-                classifier: Arc::clone(&classifier),
-                classifier_settings: settings,
-                extractor: stub_extractor(),
-                extractor_settings: ExtractorSettings::default(),
-                comprehension: stub_comprehension(),
-                comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-            },
-            PedagogyConfig::default(),
-        );
-
-        let loaded = storage
-            .load_session(session.id)
-            .await
-            .unwrap()
-            .expect("must load");
-        dm.resume_session(loaded).await.unwrap();
-
-        // Verify rehydration.
-        assert_eq!(
-            dm.learner.recent_assessments.len(),
-            1,
-            "recent_assessments must be populated from the persisted classification"
-        );
-        assert_eq!(
-            dm.learner.recent_assessments[0].state,
-            EngagementState::FrustratedTrying,
-            "rehydrated state must match what was saved"
-        );
-        assert_eq!(
-            dm.learner.current_engagement,
-            EngagementState::FrustratedTrying,
-            "current_engagement must reflect the most recent rehydrated assessment"
-        );
-    }
-
-    #[tokio::test]
-    async fn respond_to_streaming_spawns_classify_task_and_persists() {
-        use primer_classifier::{EngagementClassifier, StubEngagementClassifier};
-        use primer_core::classifier::EngagementAssessment;
-        use primer_core::storage::SessionStore;
-        use primer_storage::SqliteSessionStore;
-
-        // A classifier that always returns FrustratedTrying with high confidence.
-        let target_state = EngagementState::FrustratedTrying;
-        let classifier: Arc<dyn EngagementClassifier> = Arc::new(
-            StubEngagementClassifier::with_response(EngagementAssessment {
-                state: target_state,
-                confidence: 0.95,
-                reasoning: Some("integration test".into()),
-            }),
-        );
-
-        let storage: Arc<dyn SessionStore> = Arc::new(
-            SqliteSessionStore::open_for_locale(
-                std::path::Path::new(":memory:"),
-                primer_core::i18n::Locale::default(),
-            )
-            .unwrap(),
-        );
-
-        let backend = ScriptedBackend::new(vec![
-            Ok(chunk("Great question!", false)),
-            Ok(chunk("", true)),
-        ]);
-        let knowledge = EmptyKnowledge;
-        let settings = ClassifierSettings::default();
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&storage) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            DialogueManagerSubsystems {
-                classifier: Arc::clone(&classifier),
-                classifier_settings: settings,
-                extractor: stub_extractor(),
-                extractor_settings: ExtractorSettings::default(),
-                comprehension: stub_comprehension(),
-                comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-            },
-            PedagogyConfig::default(),
-        );
-
-        dm.open_session().await.unwrap();
-
-        // Run one full turn. After this call a classify_task should be live.
-        let response = dm
-            .respond_to_streaming("Why is the sky blue?", |_| {})
-            .await
-            .unwrap();
-        assert!(!response.is_empty(), "should have a non-empty response");
-
-        // The classify_task is now running (or already done). Simulating the
-        // start of the next turn by calling await_pending_classification
-        // should apply the FrustratedTrying assessment.
-        dm.await_pending_classification().await;
-
-        // Assessment applied: current_engagement updated by the stub.
-        assert_eq!(
-            dm.learner.current_engagement, target_state,
-            "await_pending_classification must apply the spawned assessment"
-        );
-        assert_eq!(
-            dm.learner.recent_assessments.len(),
-            1,
-            "assessment must be pushed into recent_assessments"
-        );
-    }
-
-    #[tokio::test]
-    async fn await_pending_classification_aborts_and_preserves_state_on_timeout() {
-        use primer_classifier::EngagementClassifier;
-        use primer_core::classifier::{EngagementAssessment, EngagementContext};
-        use std::time::Duration;
-
-        // Classifier that sleeps long enough to reliably exceed the test's
-        // blocking_timeout. If the timeout path works, the sleep never
-        // completes (task gets aborted) and current_engagement stays untouched.
-        struct SlowClassifier;
-
-        #[async_trait]
-        impl EngagementClassifier for SlowClassifier {
-            fn identifier(&self) -> &str {
-                "slow"
-            }
-            async fn classify(&self, _ctx: EngagementContext<'_>) -> Result<EngagementAssessment> {
-                tokio::time::sleep(Duration::from_secs(5)).await;
-                Ok(EngagementAssessment {
-                    state: EngagementState::FrustratedTrying,
-                    confidence: 0.99,
-                    reasoning: None,
-                })
-            }
-        }
-
-        let backend = ScriptedBackend::new(vec![Ok(chunk("hi", false)), Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        // Tight timeout so the await reliably trips it before the 5s sleep.
-        let settings = ClassifierSettings {
-            blocking_timeout: Duration::from_millis(50),
-            ..ClassifierSettings::default()
-        };
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend,
-            &knowledge,
-            DialogueManagerStores::default(),
-            DialogueManagerSubsystems {
-                classifier: Arc::new(SlowClassifier),
-                classifier_settings: settings,
-                extractor: stub_extractor(),
-                extractor_settings: ExtractorSettings::default(),
-                comprehension: stub_comprehension(),
-                comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-            },
-            PedagogyConfig::default(),
-        );
-
-        // Run a turn so a classify_task is spawned. The task is still
-        // sleeping when respond_to_streaming returns.
-        let _ = dm.respond_to_streaming("hi", |_| {}).await.unwrap();
-        assert!(
-            dm.classify_task.is_some(),
-            "a classify_task must be spawned after a successful turn"
-        );
-        // Capture the engagement state AFTER respond_to_streaming so the
-        // placeholder word-count heuristic in `update_learner_model` (which
-        // mutates `current_engagement` independently of the classifier) does
-        // not contaminate this test. We're checking that the timeout path
-        // does not apply the slow classifier's pending result, not that the
-        // pre-existing heuristic is bypassed.
-        let initial = dm.learner.current_engagement;
-
-        // This call should hit the timeout path: abort the task, log
-        // tracing::debug!, and return without applying any assessment.
-        let started = std::time::Instant::now();
-        dm.await_pending_classification().await;
-        let elapsed = started.elapsed();
-
-        assert!(
-            elapsed < Duration::from_secs(2),
-            "await_pending_classification must give up within ~blocking_timeout, \
-             not wait for the slow classifier; elapsed={elapsed:?}"
-        );
-        assert_eq!(
-            dm.learner.current_engagement, initial,
-            "timeout path must NOT update current_engagement"
-        );
-        assert!(
-            dm.learner.recent_assessments.is_empty(),
-            "timeout path must NOT push into recent_assessments"
-        );
-        assert!(
-            dm.classify_task.is_none(),
-            "the task handle must be consumed even on timeout"
-        );
-    }
-
-    /// Backend that serves the same single-chunk response on every
-    /// `generate_stream` call. Used by multi-turn tests where the exact
-    /// content of the Primer response does not matter.
-    struct RepeatingBackend;
-
-    #[async_trait]
-    impl InferenceBackend for RepeatingBackend {
-        fn name(&self) -> &str {
-            "repeating-test"
-        }
-        async fn is_available(&self) -> bool {
-            true
-        }
-        async fn generate_stream(
-            &self,
-            _prompt: &Prompt,
-            _params: &GenerationParams,
-        ) -> Result<TokenStream> {
-            let items: Vec<Result<TokenChunk>> = vec![Ok(chunk("ok.", false)), Ok(chunk("", true))];
-            Ok(Box::pin(stream::iter(items)))
-        }
-        async fn summarize(&self, turns: &[Turn], _target_chars: usize) -> Result<String> {
-            Ok(format!(
-                "[repeating-backend summary covering {} turns]",
-                turns.len()
-            ))
-        }
-    }
-
-    // ─── End-to-end: classifier routing across a multi-turn session ───
-
-    #[tokio::test]
-    async fn end_to_end_classifier_routing_across_multi_turn_session() {
-        use primer_classifier::{EngagementClassifier, StubEngagementClassifier};
-        use primer_core::classifier::EngagementAssessment;
-        use primer_core::conversation::PedagogicalIntent;
-        use primer_core::storage::SessionStore;
-        use primer_storage::SqliteSessionStore;
-        use std::time::Duration;
-
-        // Scripted classifier:
-        //   turn 1 -> Engaged, turn 2 -> FrustratedTrying, turn 3 -> Disengaging
-        // Exhausted script falls back to Engaged for turn 4 — but by then
-        // current_engagement is already Disengaging (applied before turn 4 starts).
-        let classifier: Arc<dyn EngagementClassifier> =
-            Arc::new(StubEngagementClassifier::with_script(vec![
-                EngagementAssessment {
-                    state: EngagementState::Engaged,
-                    confidence: 0.9,
-                    reasoning: None,
-                },
-                EngagementAssessment {
-                    state: EngagementState::FrustratedTrying,
-                    confidence: 0.9,
-                    reasoning: None,
-                },
-                EngagementAssessment {
-                    state: EngagementState::Disengaging,
-                    confidence: 0.9,
-                    reasoning: None,
-                },
-            ]));
-
-        let storage: Arc<dyn SessionStore> = Arc::new(
-            SqliteSessionStore::open_for_locale(
-                std::path::Path::new(":memory:"),
-                primer_core::i18n::Locale::default(),
-            )
-            .unwrap(),
-        );
-
-        let backend = RepeatingBackend;
-        let knowledge = EmptyKnowledge;
-
-        // Generous blocking timeout for deterministic test behaviour.
-        let settings = ClassifierSettings {
-            blocking_timeout: Duration::from_secs(5),
-            ..Default::default()
-        };
-
-        let mut learner = test_learner();
-        // 60-second threshold: a backdated session (120 s elapsed) reliably
-        // routes Disengaging → SessionClose.
-        learner.preferences.early_disengagement_threshold = Duration::from_secs(60);
-
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&storage) as Arc<dyn SessionStore>),
-                learner: None,
-            },
-            DialogueManagerSubsystems {
-                classifier: Arc::clone(&classifier),
-                classifier_settings: settings,
-                extractor: stub_extractor(),
-                extractor_settings: ExtractorSettings::default(),
-                comprehension: stub_comprehension(),
-                comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-            },
-            PedagogyConfig::default(),
-        );
-
-        dm.open_session().await.unwrap();
-
-        // Backdate started_at so elapsed (120 s) exceeds the 60-second threshold.
-        // This makes Disengaging → SessionClose rather than Encouragement.
-        dm.session.started_at = Utc::now() - chrono::Duration::seconds(120);
-
-        let session_id = dm.session.id;
-
-        // ── Turn 1 ──
-        // classify task returns Engaged (first script entry).
-        let _r1 = dm
-            .respond_to_streaming("i'm curious about gravity", |_| {})
-            .await
-            .unwrap();
-        // Drain the spawned task; apply Engaged.
-        dm.await_pending_classification().await;
-        assert_eq!(
-            dm.learner.current_engagement,
-            EngagementState::Engaged,
-            "turn 1: engagement must be Engaged"
-        );
-
-        // ── Turn 2 ──
-        // At the START of respond_to_streaming, await_pending_classification
-        // is called internally — but we already drained it above, so there is
-        // nothing to await.  After this call, a new task carrying FrustratedTrying
-        // is spawned.
-        let _r2 = dm
-            .respond_to_streaming("I think it's hard to explain", |_| {})
-            .await
-            .unwrap();
-        // Drain the spawned task; apply FrustratedTrying.
-        dm.await_pending_classification().await;
-        assert_eq!(
-            dm.learner.current_engagement,
-            EngagementState::FrustratedTrying,
-            "turn 2: engagement must be FrustratedTrying after classifier"
-        );
-
-        // ── Turn 3 ──
-        // Task for this turn returns Disengaging.
-        let _r3 = dm
-            .respond_to_streaming("I'm not sure but maybe...", |_| {})
-            .await
-            .unwrap();
-        // Drain; apply Disengaging.
-        dm.await_pending_classification().await;
-        assert_eq!(
-            dm.learner.current_engagement,
-            EngagementState::Disengaging,
-            "turn 3: engagement must be Disengaging after classifier"
-        );
-
-        // ── Turn 4 ──
-        // At the START of respond_to_streaming, await_pending_classification
-        // is called (nothing to drain — we already did it). Then decide_intent
-        // sees Disengaging + elapsed (120 s) > threshold (60 s) → SessionClose.
-        let _r4 = dm.respond_to_streaming("ok", |_| {}).await.unwrap();
-
-        // last_intent reads the intent stored on the most recent Primer turn.
-        let intent = dm.last_intent().expect("intent must be set after turn 4");
-        assert_eq!(
-            intent,
-            PedagogicalIntent::SessionClose,
-            "turn 4: Disengaging + elapsed > threshold must route to SessionClose"
-        );
-
-        // Drain the task spawned after turn 4 (not needed for intent assertion,
-        // but ensures we don't leave background work running after the test).
-        dm.await_pending_classification().await;
-
-        // All four child-turn classifications must have been persisted.
-        let recent = storage
-            .load_recent_assessments(session_id, "stub", 10)
-            .await
-            .unwrap();
-        assert_eq!(
-            recent.len(),
-            4,
-            "all four turn classifications must be persisted; got {}",
-            recent.len()
-        );
-    }
-
-    #[tokio::test]
-    async fn end_to_end_save_learner_after_open_and_one_turn() {
-        // Build a manager with both Some(SessionStore) and Some(LearnerStore)
-        // backed by the same SqliteSessionStore (which implements both traits).
-        // Run open_session + one turn and verify the learners row was upserted.
-        use primer_core::storage::LearnerStore;
-        use primer_storage::SqliteSessionStore;
-
-        let store = Arc::new(
-            SqliteSessionStore::open_for_locale(
-                std::path::Path::new(":memory:"),
-                primer_core::i18n::Locale::default(),
-            )
-            .unwrap(),
-        );
-
-        // Pre-save the learner so the DB has a row to UPDATE rather than INSERT.
-        let learner = test_learner();
-        store.save_learner(&learner).await.unwrap();
-
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hello!", false)), Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&store) as Arc<dyn SessionStore>),
-                learner: Some(Arc::clone(&store) as Arc<dyn LearnerStore>),
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _greeting = dm.open_session().await.unwrap();
-        let _reply = dm.respond_to("hello").await.unwrap();
-
-        // load_learner should return the persisted row.
-        let loaded = store
-            .load_learner()
-            .await
-            .unwrap()
-            .expect("learner row must exist");
-        assert_eq!(
-            loaded.profile.id, dm.learner.profile.id,
-            "persisted learner id must match"
-        );
-    }
-
-    #[tokio::test]
-    async fn divergence_bug_closed_via_cli_startup_flow() {
-        // Fixture: a fresh DB seeded with a session under UUID U1, no
-        // learners row yet (simulates the v3 → v4 upgrade-on-first-open).
-        // Then run the CLI's first-run startup flow:
-        //   load_learner() == None
-        //   most_recent_session_learner_id() == Some(U1)
-        //   mint LearnerModel with id=U1, save_learner(...)
-        // Assert the resulting LearnerModel.profile.id == U1.
-        use primer_core::conversation::Session as ConversationSession;
-        use primer_core::storage::{LearnerStore, SessionStore};
-        use primer_storage::SqliteSessionStore;
-
-        let store = Arc::new(
-            SqliteSessionStore::open_for_locale(
-                std::path::Path::new(":memory:"),
-                primer_core::i18n::Locale::default(),
-            )
-            .unwrap(),
-        );
-        let u1 = uuid::Uuid::new_v4();
-        let s = ConversationSession::new(u1);
-        store.save_session(&s).await.unwrap();
-
-        // Simulate the CLI startup flow.
-        let load_result = store.load_learner().await.unwrap();
-        assert!(load_result.is_none(), "no learner row yet");
-
-        let adopted = store
-            .most_recent_session_learner_id()
-            .await
-            .unwrap()
-            .expect("session exists");
-        assert_eq!(adopted, u1);
-
-        let mut adopted_learner = test_learner();
-        adopted_learner.profile.id = adopted;
-        store.save_learner(&adopted_learner).await.unwrap();
-
-        // Construct a DialogueManager with the adopted learner.
-        let backend = ScriptedBackend::new(vec![Ok(chunk("", true))]);
-        let knowledge = EmptyKnowledge;
-        let mut dm = DialogueManager::new(
-            adopted_learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: Some(Arc::clone(&store) as Arc<dyn SessionStore>),
-                learner: Some(Arc::clone(&store) as Arc<dyn LearnerStore>),
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.open_session().await.unwrap();
-
-        assert_eq!(
-            dm.session.learner_id, dm.learner.profile.id,
-            "session learner_id must match adopted learner id"
-        );
-        assert_eq!(
-            dm.session.learner_id, u1,
-            "adopted learner id must be the original session's learner_id"
-        );
-    }
-
-    // ─── Per-turn save gating (learner_dirty flag) ─────────────────────
-
-    /// Build a manager with a `CountingLearnerStore` and a learner state
-    /// that `update_learner_model` will NOT change for the chosen input.
-    /// `current_engagement = Reflecting` + a 1-or-2-word input maps to
-    /// `Reflecting` again via the `match other => other` branch, so
-    /// `current_engagement` is unchanged → no dirty → no per-turn save.
-    fn dirty_flag_test_setup(
-        starting: EngagementState,
-    ) -> (LearnerModel, Arc<CountingLearnerStore>) {
-        let mut learner = test_learner();
-        learner.current_engagement = starting;
-        let store = Arc::new(CountingLearnerStore::new());
-        (learner, store)
-    }
-
-    #[tokio::test]
-    async fn per_turn_save_skipped_when_no_persisted_field_changes() {
-        // learner starts at Reflecting; the input "ok yes" is < 3 words so
-        // update_learner_model takes the "match other => other" branch
-        // and leaves current_engagement at Reflecting. The classifier is
-        // a stub returning no assessments. The only save_learner call is
-        // the one open_session emits.
-        let (learner, store) = dirty_flag_test_setup(EngagementState::Reflecting);
-        let backend = RepeatingBackend;
-        let knowledge = EmptyKnowledge;
-
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: None,
-                learner: Some(Arc::clone(&store) as Arc<dyn LearnerStore>),
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.open_session().await.unwrap();
-        assert_eq!(
-            store.save_count(),
-            1,
-            "open_session must save once (lifecycle event)"
-        );
-
-        let _ = dm.respond_to("ok yes").await.unwrap();
-        assert_eq!(
-            store.save_count(),
-            1,
-            "per-turn save must be SKIPPED when no persisted field changed (still 1 from open)"
-        );
-    }
-
-    #[tokio::test]
-    async fn per_turn_save_fires_when_engagement_changes() {
-        // learner starts at Reflecting; a long input (>=3 words) maps to
-        // Engaged in update_learner_model, which IS a change to a
-        // persisted field. dirty=true → per-turn save fires.
-        let (learner, store) = dirty_flag_test_setup(EngagementState::Reflecting);
-        let backend = RepeatingBackend;
-        let knowledge = EmptyKnowledge;
-
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: None,
-                learner: Some(Arc::clone(&store) as Arc<dyn LearnerStore>),
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.open_session().await.unwrap();
-        let count_after_open = store.save_count();
-
-        let _ = dm
-            .respond_to("this is a longer answer with many words")
-            .await
-            .unwrap();
-        assert_eq!(
-            store.save_count(),
-            count_after_open + 1,
-            "per-turn save must fire exactly once when current_engagement changes"
-        );
-    }
-
-    #[tokio::test]
-    async fn dirty_cleared_after_save_so_subsequent_idle_turn_skips_save() {
-        // Sequence: open → dirty turn → idle turn.
-        // After the dirty turn, the flag should be cleared; the idle
-        // turn must not produce a second per-turn save.
-        let (learner, store) = dirty_flag_test_setup(EngagementState::Reflecting);
-        let backend = RepeatingBackend;
-        let knowledge = EmptyKnowledge;
-
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: None,
-                learner: Some(Arc::clone(&store) as Arc<dyn LearnerStore>),
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.open_session().await.unwrap();
-        let after_open = store.save_count();
-
-        // Dirty turn — updates engagement Reflecting → Engaged.
-        let _ = dm
-            .respond_to("this is a longer answer with many words")
-            .await
-            .unwrap();
-        let after_dirty = store.save_count();
-        assert_eq!(after_dirty, after_open + 1, "dirty turn must save");
-
-        // Idle turn — current_engagement is now Engaged, input "ok yes"
-        // (word_count<3) maps Engaged → Reflecting via the "Engaged =>
-        // Reflecting" arm, so the value DOES change. We need an input
-        // that keeps Engaged as Engaged: a long input. But that would
-        // also keep dirty stable (Engaged → Engaged is no change).
-        let _ = dm
-            .respond_to("yes that is exactly what I think")
-            .await
-            .unwrap();
-        assert_eq!(
-            store.save_count(),
-            after_dirty,
-            "idle turn (Engaged → Engaged) must NOT save again"
-        );
-    }
-
-    /// Always-failing learner store: every `save_learner` returns Err.
-    /// Used to prove that save failures are logged-and-swallowed rather
-    /// than propagated up through the dialogue-manager API.
-    struct FailingLearnerStore {
-        attempts: Mutex<u32>,
-    }
-    impl FailingLearnerStore {
-        fn new() -> Self {
-            Self {
-                attempts: Mutex::new(0),
-            }
-        }
-        fn attempt_count(&self) -> u32 {
-            *self.attempts.lock().unwrap()
-        }
-    }
-    #[async_trait]
-    impl primer_core::storage::LearnerStore for FailingLearnerStore {
-        async fn save_learner(&self, _learner: &LearnerModel) -> Result<()> {
-            *self.attempts.lock().unwrap() += 1;
-            Err(PrimerError::Storage(
-                "simulated save_learner failure".into(),
-            ))
-        }
-        async fn load_learner(&self) -> Result<Option<LearnerModel>> {
-            Ok(None)
-        }
-    }
-
-    #[tokio::test]
-    async fn save_learner_failure_does_not_propagate_through_respond_to() {
-        // A failing LearnerStore must be visible only as a tracing::warn —
-        // the conversation must continue. Otherwise a flaky disk would
-        // shut down the child's session, which is the wrong failure mode
-        // for a children's product.
-        let mut learner = test_learner();
-        learner.current_engagement = EngagementState::Reflecting;
-        let failing = Arc::new(FailingLearnerStore::new());
-        let backend = RepeatingBackend;
-        let knowledge = EmptyKnowledge;
-
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: None,
-                learner: Some(Arc::clone(&failing) as Arc<dyn LearnerStore>),
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        // open_session must succeed despite the underlying save failing.
-        let _ = dm
-            .open_session()
-            .await
-            .expect("open_session must not propagate save_learner errors");
-        let after_open = failing.attempt_count();
-        assert!(after_open >= 1, "open_session must attempt to save");
-
-        // A dirty turn must succeed despite the underlying save failing.
-        let reply = dm
-            .respond_to("this is a longer answer with many words")
-            .await
-            .expect("respond_to must not propagate save_learner errors");
-        assert!(!reply.is_empty(), "Primer reply must still come through");
-        assert!(
-            failing.attempt_count() > after_open,
-            "per-turn dirty save must be attempted, even though it errors"
-        );
-
-        // close_session must also swallow the error (no return value, no panic).
-        dm.close_session().await;
-
-        // Because every save errors, the dirty flag should still be set
-        // — the save site only clears dirty on success. This is the
-        // correct invariant: a failed save did NOT actually flush, so
-        // marking clean would be a lie.
-        assert!(
-            dm.learner_dirty,
-            "dirty must remain set when save_learner errors so a future save still runs"
-        );
-    }
-
-    #[tokio::test]
-    async fn close_session_always_saves_learner_regardless_of_dirty() {
-        // Lifecycle events flush unconditionally — they're explicit
-        // checkpoints, not "save when dirty" sites.
-        let (learner, store) = dirty_flag_test_setup(EngagementState::Engaged);
-        let backend = RepeatingBackend;
-        let knowledge = EmptyKnowledge;
-
-        let mut dm = DialogueManager::new(
-            learner,
-            &backend,
-            &knowledge,
-            DialogueManagerStores {
-                session: None,
-                learner: Some(Arc::clone(&store) as Arc<dyn LearnerStore>),
-            },
-            default_subsystems(),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.open_session().await.unwrap();
-        let after_open = store.save_count();
-        dm.close_session().await;
-        assert!(
-            store.save_count() > after_open,
-            "close_session must save unconditionally"
-        );
-    }
-
-    /// Session-store spy that records `update_turn_concepts` calls so
-    /// tests can assert the extractor's persistence side effect. Also
-    /// records `save_comprehensions` calls so chain tests can assert
-    /// the comprehension persistence side effect.
-    struct ConceptCapturingStore {
-        inner: CountingStore,
-        captures: Mutex<Vec<(usize, Vec<String>)>>,
-        comprehensions: Mutex<
-            Vec<(
-                usize,
-                Vec<primer_core::comprehension::ComprehensionAssessment>,
-                String,
-            )>,
-        >,
-    }
-
-    impl ConceptCapturingStore {
-        fn new() -> Self {
-            Self {
-                inner: CountingStore::new(),
-                captures: Mutex::new(vec![]),
-                comprehensions: Mutex::new(vec![]),
-            }
-        }
-        fn captured(&self) -> Vec<(usize, Vec<String>)> {
-            self.captures.lock().unwrap().clone()
-        }
-        fn captured_comprehensions(
-            &self,
-        ) -> Vec<(
-            usize,
-            Vec<primer_core::comprehension::ComprehensionAssessment>,
-            String,
-        )> {
-            self.comprehensions.lock().unwrap().clone()
-        }
-    }
-
-    #[async_trait]
-    impl primer_core::storage::SessionStore for ConceptCapturingStore {
-        async fn save_session(&self, session: &Session) -> Result<()> {
-            self.inner.save_session(session).await
-        }
-        async fn load_session(&self, id: uuid::Uuid) -> Result<Option<Session>> {
-            self.inner.load_session(id).await
-        }
-        async fn retrieve_session_turns(
-            &self,
-            session_id: uuid::Uuid,
-            query: &str,
-            k: usize,
-            exclude_indices_at_or_after: usize,
-        ) -> Result<Vec<Turn>> {
-            self.inner
-                .retrieve_session_turns(session_id, query, k, exclude_indices_at_or_after)
-                .await
-        }
-        async fn save_classification(
-            &self,
-            session_id: primer_core::conversation::SessionId,
-            turn_index: usize,
-            assessment: &primer_core::classifier::EngagementAssessment,
-            classifier_identifier: &str,
-        ) -> Result<()> {
-            self.inner
-                .save_classification(session_id, turn_index, assessment, classifier_identifier)
-                .await
-        }
-        async fn load_recent_assessments(
-            &self,
-            session_id: primer_core::conversation::SessionId,
-            classifier_identifier: &str,
-            k: usize,
-        ) -> Result<Vec<primer_core::classifier::EngagementAssessment>> {
-            self.inner
-                .load_recent_assessments(session_id, classifier_identifier, k)
-                .await
-        }
-        async fn most_recent_session_learner_id(&self) -> Result<Option<uuid::Uuid>> {
-            self.inner.most_recent_session_learner_id().await
-        }
-        async fn update_turn_concepts(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            turn_index: usize,
-            concepts: &[String],
-        ) -> Result<()> {
-            self.captures
-                .lock()
-                .unwrap()
-                .push((turn_index, concepts.to_vec()));
-            Ok(())
-        }
-
-        async fn update_exchange_concepts(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            child_turn_index: usize,
-            child_concepts: &[String],
-            primer_turn_index: usize,
-            primer_concepts: &[String],
-        ) -> Result<()> {
-            // Mirror the storage impl: push only the speaker(s) that
-            // actually have concepts, so tests that scripted an
-            // empty-on-one-side extraction don't see a phantom capture.
-            let mut captures = self.captures.lock().unwrap();
-            if !child_concepts.is_empty() {
-                captures.push((child_turn_index, child_concepts.to_vec()));
-            }
-            if !primer_concepts.is_empty() {
-                captures.push((primer_turn_index, primer_concepts.to_vec()));
-            }
-            Ok(())
-        }
-
-        async fn save_comprehensions(
-            &self,
-            _session_id: primer_core::conversation::SessionId,
-            primer_turn_index: usize,
-            assessments: &[primer_core::comprehension::ComprehensionAssessment],
-            classifier_identifier: &str,
-        ) -> Result<()> {
-            self.comprehensions.lock().unwrap().push((
-                primer_turn_index,
-                assessments.to_vec(),
-                classifier_identifier.to_string(),
-            ));
-            Ok(())
-        }
-    }
-
-    #[tokio::test]
-    async fn extract_task_persists_concepts_for_both_turns_after_response() {
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hi there!", true))]);
-        let extractor = Arc::new(primer_extractor::StubConceptExtractor::with_response(
-            ConceptExtraction {
-                child_concepts: vec!["topic-a".into()],
-                primer_concepts: vec!["topic-b".into()],
-            },
-        ));
-        let store = Arc::new(ConceptCapturingStore::new());
-
-        let stores = DialogueManagerStores {
-            session: Some(store.clone() as Arc<dyn primer_core::storage::SessionStore>),
-            learner: None,
-        };
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend as &dyn InferenceBackend,
-            &EmptyKnowledge as &dyn KnowledgeBase,
-            stores,
-            subsystems_with_extractor(extractor as Arc<dyn ConceptExtractor>),
-            PedagogyConfig::default(),
-        );
-
-        dm.respond_to("Hello").await.unwrap();
-
-        // Yield until the spawned extractor task lands its captures.
-        for _ in 0..50 {
-            if store.captured().len() >= 2 {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-        }
-
-        let captures = store.captured();
-        assert_eq!(captures.len(), 2, "expected child + primer captures");
-        // Child turn is at index 0, primer at index 1.
-        let child_capture = captures.iter().find(|(i, _)| *i == 0).unwrap();
-        let primer_capture = captures.iter().find(|(i, _)| *i == 1).unwrap();
-        assert_eq!(child_capture.1, vec!["topic-a".to_string()]);
-        assert_eq!(primer_capture.1, vec!["topic-b".to_string()]);
-    }
-
-    #[tokio::test]
-    async fn extract_task_does_not_spawn_on_inference_error() {
-        let backend = ScriptedBackend::new(vec![Err(PrimerError::Inference("boom".into()))]);
-        let extractor = Arc::new(primer_extractor::StubConceptExtractor::with_response(
-            ConceptExtraction {
-                child_concepts: vec!["should-not-persist".into()],
-                primer_concepts: vec![],
-            },
-        ));
-        let store = Arc::new(ConceptCapturingStore::new());
-
-        let stores = DialogueManagerStores {
-            session: Some(store.clone() as Arc<dyn primer_core::storage::SessionStore>),
-            learner: None,
-        };
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend as &dyn InferenceBackend,
-            &EmptyKnowledge as &dyn KnowledgeBase,
-            stores,
-            subsystems_with_extractor(extractor as Arc<dyn ConceptExtractor>),
-            PedagogyConfig::default(),
-        );
-
-        let _ = dm.respond_to("Hello").await;
-
-        // Give the runtime a chance to run any spuriously-spawned task.
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        assert!(
-            store.captured().is_empty(),
-            "extractor must not run on inference error"
-        );
-    }
-
-    #[tokio::test]
-    async fn pending_extraction_applied_to_learner_at_next_turn() {
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hi turn 1!", true))]);
-        // Two turns of extraction scripted: turn 1 surfaces "gravity" + "physics",
-        // turn 2 surfaces "mass". Only the first one matters for this test —
-        // we want to assert that after respond_to(turn 2), the learner has
-        // gravity from turn 1's extraction.
-        let extractor = Arc::new(primer_extractor::StubConceptExtractor::with_script(vec![
-            ConceptExtraction {
-                child_concepts: vec!["gravity".into()],
-                primer_concepts: vec!["physics".into()],
-            },
-            ConceptExtraction {
-                child_concepts: vec!["mass".into()],
-                primer_concepts: vec![],
-            },
-        ]));
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend as &dyn InferenceBackend,
-            &EmptyKnowledge as &dyn KnowledgeBase,
-            DialogueManagerStores::default(),
-            subsystems_with_extractor(extractor as Arc<dyn ConceptExtractor>),
-            PedagogyConfig::default(),
-        );
-
-        dm.respond_to("turn 1").await.unwrap();
-
-        // Refill the backend script for turn 2.
-        backend.set_script(vec![Ok(chunk("Hi turn 2!", true))]);
-
-        // Allow the previous-turn extractor task to complete before turn 2 starts.
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-
-        dm.respond_to("turn 2").await.unwrap();
-
-        let names: std::collections::HashSet<&str> = dm
-            .learner
-            .concepts
-            .iter()
-            .map(|c| c.concept_id.as_str())
-            .collect();
-        assert!(
-            names.contains("gravity"),
-            "child concept 'gravity' should be applied to learner; got: {:?}",
-            names
-        );
-        assert!(
-            names.contains("physics"),
-            "primer concept 'physics' should be applied to learner; got: {:?}",
-            names
-        );
-    }
-
-    #[tokio::test]
-    async fn close_session_drains_extractor_task() {
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hi", true))]);
-        let extractor = Arc::new(primer_extractor::StubConceptExtractor::with_response(
-            ConceptExtraction {
-                child_concepts: vec!["x".into()],
-                primer_concepts: vec![],
-            },
-        ));
-        let store = Arc::new(ConceptCapturingStore::new());
-
-        let stores = DialogueManagerStores {
-            session: Some(store.clone() as Arc<dyn primer_core::storage::SessionStore>),
-            learner: None,
-        };
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend as &dyn InferenceBackend,
-            &EmptyKnowledge as &dyn KnowledgeBase,
-            stores,
-            subsystems_with_extractor(extractor as Arc<dyn ConceptExtractor>),
-            PedagogyConfig::default(),
-        );
-
-        dm.respond_to("hi").await.unwrap();
-        // close_session must drain so the extractor's update_turn_concepts
-        // call has landed by the time close returns.
-        dm.close_session().await;
-
-        let captures = store.captured();
-        assert!(
-            !captures.is_empty(),
-            "expected extraction to land before close returns"
-        );
-    }
-
-    // ─── Chained post-response (extraction → comprehension) ──────────
-
-    #[tokio::test]
-    async fn post_response_chain_persists_extraction_and_comprehension() {
-        use primer_comprehension::StubComprehensionClassifier;
-        use primer_core::comprehension::{ComprehensionAssessment, ComprehensionResult};
-        use primer_core::extractor::ConceptExtraction;
-        use primer_core::learner::UnderstandingDepth;
-
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hi there!", true))]);
-        let extractor = Arc::new(primer_extractor::StubConceptExtractor::with_response(
-            ConceptExtraction {
-                child_concepts: vec!["gravity".into()],
-                primer_concepts: vec![],
-            },
-        ));
-        let comprehension = Arc::new(StubComprehensionClassifier::with_response(
-            ComprehensionResult {
-                assessments: vec![ComprehensionAssessment {
-                    concept: "gravity".into(),
-                    depth: UnderstandingDepth::Recall,
-                    confidence: 0.8,
-                    evidence: Some("named the concept".into()),
-                }],
-            },
-        )) as Arc<dyn primer_comprehension::ComprehensionClassifier>;
-        let store = Arc::new(ConceptCapturingStore::new());
-
-        let stores = DialogueManagerStores {
-            session: Some(store.clone() as Arc<dyn primer_core::storage::SessionStore>),
-            learner: None,
-        };
-
-        // Build subsystems directly (subsystems_with_extractor doesn't
-        // accept a custom comprehension classifier).
-        let subsystems = DialogueManagerSubsystems {
-            classifier: stub_classifier(),
-            classifier_settings: ClassifierSettings::default(),
-            extractor: extractor as Arc<dyn ConceptExtractor>,
-            extractor_settings: ExtractorSettings::default(),
-            comprehension,
-            comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-        };
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend as &dyn InferenceBackend,
-            &EmptyKnowledge as &dyn KnowledgeBase,
-            stores,
-            subsystems,
-            PedagogyConfig::default(),
-        );
-
-        dm.respond_to("Hello").await.unwrap();
-        // close_session drains the post-response chain so both
-        // update_exchange_concepts and save_comprehensions have landed
-        // and in-memory state has been applied by the time it returns.
-        dm.close_session().await;
-
-        // Extraction persisted: child concept captured.
-        let captures = store.captured();
-        assert!(
-            captures
-                .iter()
-                .any(|(_, names)| names.contains(&"gravity".to_string())),
-            "expected child capture of 'gravity'; got {:?}",
-            captures
-        );
-
-        // Comprehension persisted via save_comprehensions.
-        let comp_captures = store.captured_comprehensions();
-        assert_eq!(
-            comp_captures.len(),
-            1,
-            "expected one save_comprehensions call; got {:?}",
-            comp_captures
-        );
-        let (primer_idx, assessments, classifier_id) = &comp_captures[0];
-        assert_eq!(*primer_idx, 1, "primer turn index should be 1");
-        assert_eq!(assessments.len(), 1);
-        assert_eq!(assessments[0].concept, "gravity");
-        assert_eq!(assessments[0].depth, UnderstandingDepth::Recall);
-        assert_eq!(classifier_id, "stub");
-
-        // Last comprehension applied to learner via await_pending_background.
-        let last_comp = dm
-            .last_comprehension()
-            .expect("last_comprehension must be set after the chain runs");
-        assert_eq!(last_comp.assessments.len(), 1);
-        assert_eq!(last_comp.assessments[0].concept, "gravity");
-
-        // learner.concepts has gravity at Recall (extraction inserted at
-        // Aware first; comprehension promoted to Recall via monotonic max).
-        let gravity = dm
-            .learner
-            .concepts
-            .iter()
-            .find(|c| c.concept_id == "gravity")
-            .expect("'gravity' must be in learner concepts");
-        assert_eq!(gravity.depth, UnderstandingDepth::Recall);
-    }
-
-    #[tokio::test]
-    async fn post_response_chain_skips_comprehension_on_empty_extraction() {
-        // Stub extractor returns empty → candidate_concepts is empty →
-        // comprehension MUST NOT be invoked. The spy comprehension
-        // classifier panics if classify() is called.
-        struct PanicOnCall;
-        #[async_trait]
-        impl primer_comprehension::ComprehensionClassifier for PanicOnCall {
-            fn identifier(&self) -> &str {
-                "panic"
-            }
-            async fn classify(
-                &self,
-                _ctx: primer_core::comprehension::ComprehensionContext<'_>,
-            ) -> Result<primer_core::comprehension::ComprehensionResult> {
-                panic!("comprehension must not be called when extractor returned empty");
-            }
-        }
-
-        let backend = ScriptedBackend::new(vec![Ok(chunk("Hi there!", true))]);
-        let extractor =
-            Arc::new(primer_extractor::StubConceptExtractor::new()) as Arc<dyn ConceptExtractor>;
-        let comprehension =
-            Arc::new(PanicOnCall) as Arc<dyn primer_comprehension::ComprehensionClassifier>;
-        let store = Arc::new(ConceptCapturingStore::new());
-
-        let stores = DialogueManagerStores {
-            session: Some(store.clone() as Arc<dyn primer_core::storage::SessionStore>),
-            learner: None,
-        };
-
-        let subsystems = DialogueManagerSubsystems {
-            classifier: stub_classifier(),
-            classifier_settings: ClassifierSettings::default(),
-            extractor,
-            extractor_settings: ExtractorSettings::default(),
-            comprehension,
-            comprehension_settings: primer_comprehension::ComprehensionSettings::default(),
-        };
-
-        let mut dm = DialogueManager::new(
-            test_learner(),
-            &backend as &dyn InferenceBackend,
-            &EmptyKnowledge as &dyn KnowledgeBase,
-            stores,
-            subsystems,
-            PedagogyConfig::default(),
-        );
-
-        dm.respond_to("Hello").await.unwrap();
-        // Drain. If PanicOnCall.classify() were called, the panicked
-        // task would surface as an Err inside apply_post_response_outcome
-        // (logged) — but the comprehension code path guards on
-        // candidates.is_empty() before invoking classify, so classify
-        // is never reached and no panic surfaces.
-        dm.close_session().await;
-
-        // No comprehension captures because classify was never called.
-        let comp_captures = store.captured_comprehensions();
-        assert!(
-            comp_captures.is_empty(),
-            "save_comprehensions must not be invoked when extraction is empty; got {:?}",
-            comp_captures
-        );
-    }
+    mod background_tests;
+    mod lifecycle_tests;
+    mod turn_tests;
 }
